requirements.txt
streamlit
pandas
tensorflow
Pillow
matplotlib
seaborn
scikit-learn